pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iescan"
version = "0.1.0"
description = "Internal Enterprise Security Assessment and Network scanner for authorized penetration testing"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "ldap3>=2.9",
    "impacket>=0.11",
    "pyasn1>=0.5",
    "dnspython>=2.4",
    "rich>=13.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "mcp>=1.0",
    "httpx>=0.27",
    "pysmb>=1.2",
    "pymssql>=2.2",
    "cx_Oracle>=8.3",
    "psycopg2-binary>=2.9",
    "scapy>=2.5",
    "pydantic>=2.0",
    "cryptography>=41.0",
    "ipaddress>=1.0.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
iescan = "iescan.cli:main"

[project.entry-points."mcp.servers"]
iescan = "iescan.mcp.server:serve"

[tool.hatch.build.targets.wheel]
packages = ["src/iescan"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
